+++ packages/widgetbook/lib/src/next/component.dart
@@ -1,8 +1,12 @@
+import 'package:flutter/widgets.dart';
+
 import '../navigation/nodes/nodes.dart' as v3;
+import 'args/story_args.dart';
 import 'meta.dart';
 import 'story.dart';
 
-class Component<T> extends v3.WidgetbookComponent {
+class Component<TWidget extends Widget, TArgs extends StoryArgs<TWidget>>
+    extends v3.WidgetbookComponent {
   Component({
     required this.meta,
     required this.stories,
@@ -11,6 +15,6 @@ class Component<T> extends v3.WidgetbookComponent {
           useCases: stories,
         );
 
-  final Meta<T> meta;
-  final List<Story<T>> stories;
+  final Meta<TWidget> meta;
+  final List<Story<TWidget, TArgs>> stories;
 }