+++ packages/widgetbook/lib/src/next/meta.dart
@@ -1,7 +1,14 @@
+import 'args/story_args.dart';
+
 class Meta<T> {
-  Meta({
+  const Meta({
     String? name,
-  }) : name = name == null ? T.toString() : name;
+  }) : name = name ?? '$T';
 
   final String name;
+}
+
+/// Same as [Meta] but for custom [StoryArgs].
+class MetaWithArgs<TWidget, TArgs> extends Meta<TWidget> {
+  const MetaWithArgs({super.name});
 }