+++ packages/widgetbook/lib/src/next/story.dart
@@ -1,37 +1,44 @@
 import 'package:flutter/material.dart';
 
-import '../fields/fields.dart';
 import '../navigation/nodes/nodes.dart' as v3;
-import '../state/state.dart';
 import 'args/story_args.dart';
 
-typedef StoryBuilder = Widget Function(BuildContext context, Widget story);
+typedef SetupBuilder = Widget Function(BuildContext context, Widget story);
+typedef ArgsBuilder<TWidget extends Widget, TArgs extends StoryArgs<TWidget>>
+    = TWidget Function(BuildContext context, TArgs args);
 
 @optionalTypeArgs
-class Story<TWidget> extends v3.WidgetbookUseCase {
+abstract class Story<TWidget extends Widget, TArgs extends StoryArgs<TWidget>>
+    extends v3.WidgetbookUseCase {
   Story({
     required super.name,
-    required this.args,
     super.designLink,
-    this.setup,
+    this.setup = defaultSetup,
+    required this.args,
+    required this.argsBuilder,
   }) : super(
-          builder: (context) {
-            final state = WidgetbookState.of(context);
-            final groupMap = FieldCodec.decodeQueryGroup(
-              state.queryParams['args'],
-            );
-
-            final story = args.build(context, groupMap);
-
-            return setup != null ? setup(context, story) : story;
-          },
+          builder: (context) => const SizedBox.shrink(), // TODO: remove
         );
 
-  final StoryArgs<TWidget> args;
-  final StoryBuilder? setup;
+  final TArgs args;
+  final SetupBuilder setup;
+  final ArgsBuilder<TWidget, TArgs> argsBuilder;
+
+  static Widget defaultSetup(
+    BuildContext context,
+    Widget story,
+  ) {
+    return story;
+  }
 
   @override
-  Story<TWidget> copyWith({
+  Widget build(BuildContext context) {
+    final story = argsBuilder(context, args);
+    return setup(context, story);
+  }
+
+  @override
+  Story<TWidget, TArgs> copyWith({
     String? name,
     List<v3.WidgetbookNode>? children,
   }) {