+++ packages/widgetbook_generator/lib/src/next/args_class_builder.dart
@@ -7,44 +7,30 @@ import 'arg_builder.dart';
 import 'extensions.dart';
 
 class ArgsClassBuilder {
-  ArgsClassBuilder(this.type);
+  ArgsClassBuilder(this.widgetType, this.argsType);
 
-  final DartType type;
+  final DartType widgetType;
+  final DartType argsType;
 
   Iterable<ParameterElement> get params {
-    return (type.element as ClassElement)
+    return (argsType.element as ClassElement)
         .constructors
         .first
         .parameters
         .whereNot((param) => param.name == 'key');
   }
 
-  InvokeExpression instantiate(
-    Expression Function(ParameterElement) assigner,
-  ) {
-    return InvokeExpression.newOf(
-      refer(type.displayName),
-      params //
-          .where((param) => param.isPositional)
-          .map(assigner)
-          .toList(),
-      params //
-          .where((param) => param.isNamed)
-          .lastBy((param) => param.name)
-          .map(
-            (_, param) => MapEntry(
-              param.name,
-              assigner(param),
-            ),
-          ),
-    );
-  }
-
   Class build() {
     return Class(
       (b) => b
-        ..name = '${type.displayName}Args'
-        ..extend = refer('StoryArgs<${type.displayName}>')
+        ..name = '${argsType.displayName}Args'
+        ..extend = TypeReference(
+          (b) => b
+            ..symbol = 'StoryArgs'
+            ..types.addAll([
+              refer(widgetType.displayName),
+            ]),
+        )
         ..fields.addAll(
           params.map(
             (param) => ArgBuilder(param).buildField(),
@@ -94,49 +80,20 @@ class ArgsClassBuilder {
               ),
           ),
         ])
-        ..methods.addAll(
-          [
-            Method(
-              (b) => b
-                ..name = 'list'
-                ..type = MethodType.getter
-                ..returns = refer('List<Arg>')
-                ..annotations.add(refer('override'))
-                ..lambda = true
-                ..body = literalList(
-                  params.map(
-                    (param) => refer(param.name),
-                  ),
-                ).code,
-            ),
-            Method(
-              (b) => b
-                ..name = 'build'
-                ..annotations.add(refer('override'))
-                ..returns = refer('Widget')
-                ..requiredParameters.add(
-                  Parameter(
-                    (b) => b
-                      ..name = 'context'
-                      ..type = refer('BuildContext'),
-                  ),
-                )
-                ..optionalParameters.add(
-                  Parameter(
-                    (b) => b
-                      ..name = 'group'
-                      ..type = refer('Map<String, String>?'),
-                  ),
-                )
-                ..body = instantiate(
-                  (param) => refer('resolve').call([
-                    refer(param.name),
-                    refer('context'),
-                    refer('group'),
-                  ]),
-                ).returned.statement,
-            ),
-          ],
+        ..methods.add(
+          Method(
+            (b) => b
+              ..name = 'list'
+              ..type = MethodType.getter
+              ..returns = refer('List<Arg>')
+              ..annotations.add(refer('override'))
+              ..lambda = true
+              ..body = literalList(
+                params.map(
+                  (param) => refer(param.name),
+                ),
+              ).code,
+          ),
         ),
     );
   }